+++ packages/editor/editor-mobile-bridge/src/editor/web-to-native/ios-impl.ts
@@ -1,4 +1,4 @@
-import { Color as StatusColor } from '@atlaskit/status/element';
+import { Color as StatusColor } from '@atlaskit/status';
 import { EditorBridges, EditorPluginBridges } from './index';
 import NativeBridge from './bridge';
 import { sendToBridge } from '../../bridge-utils';
@@ -111,12 +111,23 @@ export default class IosBridge implements NativeBridge {
     }
   }
 
-  currentSelection(text: string, url: string) {
+  currentSelection(
+    text: string,
+    url: string,
+    top: number,
+    right: number,
+    bottom: number,
+    left: number,
+  ) {
     if (window.webkit && window.webkit.messageHandlers.linkBridge) {
       window.webkit.messageHandlers.linkBridge.postMessage({
         name: 'currentSelection',
         text,
         url,
+        top,
+        right,
+        bottom,
+        left,
       });
     }
   }