+++ packages/editor/renderer/__tests-karma__/browser/react/nodes/status.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
-import { Status } from '@atlaskit/status/element';
+import { Status } from '@atlaskit/status';
 import StatusNode from '../../../../src/react/nodes/status';
 
 describe('Renderer - React/Nodes/Status', () => {