+++ packages/editor/renderer/src/react/nodes/status.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { PureComponent } from 'react';
-import { Status as AkStatus, Color } from '@atlaskit/status/element';
+import { Status as AkStatus, Color } from '@atlaskit/status';
 import { FabricElementsAnalyticsContext } from '@atlaskit/analytics-namespaced-context';
 
 export interface Props {