+++ packages/jira/refinement-bar/src/components/index.js
@@ -1,0 +1,8 @@
+// @flow
+
+export {
+  RefinementBarProvider,
+  RefinementBarConsumer,
+  RefinementBarContext,
+} from './ContextProvider';
+export { RefinementBarUI, RefinementBar as default } from './RefinementBar';