+++ packages/jira/refinement-bar/src/fields/Search/Controller.js
@@ -1,0 +1,7 @@
+// @flow
+
+import FieldController from '../Controller';
+
+export default class SearchController extends FieldController {
+  getInitialValue = () => '';
+}